// src/app/pokemon/interfaces/pokemon.interface.ts
export interface favorite {
  name: string;
  alias: string;
  createdAt: Date;
  isFavorite: Boolean
}

export interface RepPokemonList {
  count:    number;
  next:     string;
  previous: null;
  results:  Result[];
}


export interface PokemonDetail {
  sprites: {
    front_default: string;
  }
}

export interface Result {
  name: string;
  url:  string;
}